+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Version 64
 * 🔨 Do not crash if permission to set alarms and reminders has been removed, schedule inexact
   episode notifications instead.
 * 🔧 Improve when the app wakes the device to notify about upcoming episodes.
+* 📝 Latest user interface translations from Crowdin.
 
 #### 64.0.4*
 *2022-05-06*